Always load data to "now" (#4)
Previously after the initial load the daily load would only request data for the 24 hours after tha most recent entry in the series. That works most of the time but when the container is stopped (crash or external issue) it doesn't catch up when restarted, but required manually restarting until all missed data was requested one day at a time. This change updates the to_date on the request to "now" so missed data is retrieved on the next load.